Big clean up
-CSS sorted
-console logs removed
-unnecessary comments cleaned up
-slices moved into own folder structure under app
-eslint warnings rectified
