+++ server/controllers/projects.js
@@ -53,8 +53,6 @@ exports.getProject = function (req, res, next) {
     
         readStream.on('end', ()=>{
             const gridData = JSON.parse(gridBuffer.toString());
-            console.log("done downloading: ", gridData);    
-            // clean up memory
             gridBuffer = null; 
             readStream.destroy();  
             return res.status(200).send({project, gridData});
@@ -62,14 +60,12 @@ exports.getProject = function (req, res, next) {
     })      
 }
 exports.deleteProject = function (req,res,next){
-    console.log("delete project", req.body)
     Project.deleteOne({_id: req.body.id}, function (err){
         if (err) {return next(err)}
         return res.status(200).send({status: 'ok', id:req.body.id})
     })
 }
 exports.saveProject = function (req, res, next) {
-    console.log("request", req.body)
     const projectColors = req.body.projectColors;
     const rows = req.body.rows;
     const columns = req.body.columns;
@@ -87,12 +83,9 @@ exports.saveProject = function (req, res, next) {
 
     gridDataStream.pipe(gridFSBucket.openUploadStream(`projectGrid-${projectId}.txt`))
     .on('error', (error)=>{
-        console.log("GridFS upload error:"+ error);
         return res.send(error);
     })
-    .on('finish', (data)=>{
-        console.log("done uploading: ", data);
-        
+    .on('finish', (data)=>{      
         Project.findOneAndUpdate(
             { _id: projectId }, 
             { $set: {
