+++ src/app/slices/colorSelectorSlice.js
