+++ src/features/Builder/Builder.scss
@@ -1,48 +1,48 @@
 .Builder__work-area {
-  height: 65vh;
-  width: 100%;
   background-color: #e5e5ea;
+  height: 65vh;
   overflow: auto;
+  width: 100%;
 }
 .Builder__tool-area {
   background-color: #fffffff2;
-  padding: 0 20px;
   bottom: 0;
   left: 0;
-  width: 97%;
   margin-top: 10px;
+  padding: 0 20px;
+  width: 97%;
 }
 
 .Builder__tool-container {
-  position: relative;
+  align-items: flex-start;
   display: flex;
   flex-direction: row;
-  align-items: flex-start;
   justify-content: space-between;
+  position: relative;
   width: 100%;
 }
 .Builder__header{
+  align-items: center;
   display: flex;
   flex-direction: row;
   justify-content: space-between;
-  align-items: center;
   padding: 10px;
   width: 98%;
 }
 
 .Builder__tool_toggle {
-  font-size: 2em;
-  background-color: #fffffff2;
-  color: black;
-  width: 40px;
-  height: 40px;
-  display: flex;
-  padding: 5px;
-  justify-content: flex-start;
   align-items: center;
-  position: absolute;
-  left: -50px;
-  top: 0;
+  background-color: #fffffff2;
   bottom: 0;
+  color: black;
+  display: flex;
+  font-size: 2em;
+  height: 40px;
+  justify-content: flex-start;
+  left: -50px;
   margin: auto;
+  padding: 5px;
+  position: absolute;
+  top: 0;
+  width: 40px;
 }
