+++ src/features/Builder/components/Cell/Cell.scss
@@ -1,25 +1,25 @@
 .Cell {
-  width: 30px;
+  cursor: pointer;
+  display: flex;
   height: 30px;
   margin: 0;
-  display: flex;
-  position: relative;
   overflow: hidden;
-  cursor: pointer;
+  position: relative;
+  width: 30px;
 }
 .Cell:after {
-  width: 0;
-  height: 0;
+  border-color: #007bff transparent transparent transparent;
   border-style: solid;
   border-width: 20px 20px 0 20px;
-  border-color: #007bff transparent transparent transparent;
+  height: 0;
+  width: 0;
 }
 .Cell__x {
-  width: 100%;
-  height: 100%;
-  font-size: 18em;
-  display: flex;
-  justify-content: center;
   align-items: center;
+  display: flex;
+  font-size: 18em;
+  height: 100%;
+  justify-content: center;
   transform: translateY(-0.17em);
+  width: 100%;
 }
