+++ src/features/Builder/components/ColorSelector/ColorSelector.scss
@@ -1,9 +1,9 @@
 .ColorSelector{
-    position: relative;
+    align-items: center;
     display: flex;
     flex-direction: row;
     justify-content: flex-start;
-    align-items: center;
+    position: relative;
 }
 
 .ColorSelector__picker{
@@ -11,29 +11,29 @@
    
 }
 .ColorSelector__current-color{
-    height: 80px;
-    width: 100px;
-    margin-right: 2px;
-    border-radius: 5px;
-    margin-top: 2px;
     border: 1px dashed #545050;
-    margin-bottom: 1em;
+    border-radius: 5px;
     cursor: pointer;
+    height: 80px;
+    margin-bottom: 1em;
+    margin-right: 2px;
+    margin-top: 2px;
+    width: 100px;
 }
 .ColorSelector__display{
+    align-items: center;
     display: flex;
     flex-direction: column;
     justify-content: flex-start;
-    align-items: center;
     margin-left: 1em;
     width: 110px;
     input{
-        text-align: center;
+        border: 1px solid gainsboro;
+        border-radius: 5px;
         font-family: helvetica;
         font-size: 1.2em;
+        text-align: center;
         width: 88%;
-        border-radius: 5px;
-        border: 1px solid gainsboro;
 
         &:focus{
             border: 1px solid gainsboro;
@@ -41,14 +41,14 @@
     }
 }
 .react-colorful {
-    position: relative;
+    cursor: default;
     display: flex;
     flex-direction: column;
-    width: 300px !important;
     height: 150px !important;
+    position: relative;
     -webkit-user-select: none;
     -moz-user-select: none;
     -ms-user-select: none;
     user-select: none;
-    cursor: default;
+    width: 300px !important;
 }
