+++ src/features/Builder/components/GridContainer/GridContainer.scss
@@ -2,10 +2,9 @@
     border: 1px solid gainsboro;
 }
 .GridContainer{
+    column-gap: 10px;
     display: grid;
     grid-template-columns: repeat(50, 20px);
     grid-template-rows: repeat(50, 15px);
-    column-gap: 10px;
-    row-gap: 15px;
-    
+    row-gap: 15px;    
 }
