+++ src/features/Builder/components/GridTools/GridTools.scss
@@ -3,8 +3,8 @@
 }
 
 .GridTools__button{
-    width: 50px;
     height: 50px;
+    width: 50px;
 }
 
 .GridTools__button--selected{
