+++ src/features/Builder/components/ProjectColors/ProjectColors.scss
@@ -1,46 +1,46 @@
 .ProjectColors{
-    width: 150px;
     padding: 0px 10px 10px 10px;
+    width: 150px;
     h3{
+    font-size: 1rem;
       
     margin-top: 0;
-    font-size: 1rem;
 
     }
 }
 
 .ProjectColors__color{
-    height: 30px;
-    width: 30px;
-    margin-right: 2px;
-    border-radius: 5px;
-    margin-top: 2px;
     border: 1px dashed #545050;
+    border-radius: 5px;
+    height: 30px;
+    margin-right: 2px;
+    margin-top: 2px;
+    width: 30px;
 }
 
 .ProjectColor__colors{
+    align-items: flex-start;
     display: flex;
     flex-direction: row;
     flex-wrap: wrap;
     justify-content: flex-start;
-    align-items: flex-start;
     button {
-        height: 32px;
-        width: 32px;
-        margin-right: 2px;
-        border-radius: 5px;
-        margin-top: 2px;
-        border: 2px dashed gray;
-        background: white;
-        color: gray;
-        font-size: 2.5em;
-        display: flex;
-        justify-content: center;
         align-items: center;
+        background: white;
+        border: 2px dashed gray;
+        border-radius: 5px;
+        color: gray;
         cursor: pointer;
+        display: flex;
+        font-size: 2.5em;
+        height: 32px;
+        justify-content: center;
+        margin-right: 2px;
+        margin-top: 2px;
+        width: 32px;
     }
 
 }
 .ProjectColors__color--selected{
-    border: 1px dotted gainsboro
+    border: 1px dotted gainsboro;
 }
