+++ src/features/Builder/components/RecentColors/index.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { useSelector, useDispatch } from 'react-redux';
-import { addToProjectColors } from 'projectSlice';
+import { addToProjectColors } from 'app/slices/projectSlice';
 import './RecentColors.scss';
 
 const RecentColors = () => {
