+++ src/features/Header/Header.scss
@@ -1,15 +1,14 @@
 .Header{
-    height: 5vh;
-    min-height: 50px;
+    align-items: center;
     background: #ffffff;
-    width: 80%;
-    padding: 5px 10%;
+    border-bottom: 1px solid gainsboro;
     display: flex;
     flex-direction: row;
+    height: 5vh;
     justify-content: space-between;
-    align-items: center;
-    border-bottom: 1px solid gainsboro;
-
+    min-height: 50px;
+    padding: 5px 10%;
+    width: 80%;
 }
 
 .Header__logo{
