+++ src/pages/Dashboard/Dashboard.scss
@@ -1,8 +1,8 @@
 .Dashboard {
-  width: 80%;
-  margin: auto;
+  align-items: flex-start;
   display: flex;
   flex-direction: column;
   justify-content: flex-start;
-  align-items: flex-start;
+  margin: auto;
+  width: 80%;
 }
