+++ src/pages/Dashboard/components/CreateTile/CreateTile.scss
@@ -1,21 +1,21 @@
 .CreateTile {
-  width: 229px;
-  height: 244px;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
   align-items: center;
   border: 4px dashed gainsboro;
   border-radius: 1rem;
+  display: flex;
+  flex-direction: column;
+  height: 244px;
+  justify-content: center;
   margin: 0.5rem;
+  width: 229px;
 }
 .CreateTile__add {
   border: 2px dashed gainsboro;
   border-radius: 0.3em;
-  font-size: 4em;
-  width: 1.3em;
   color: gray;
   cursor: pointer;
+  font-size: 4em;
+  width: 1.3em;
 }
 .CreateTile__title {
   color: #5a5a5a;
@@ -23,11 +23,11 @@
 
 .CreateTile__input {
   border: none;
-  text-align: center;
-  margin-bottom: 1em;
   font-family: sans-serif;
   font-size: 1.1em;
-  max-width: 90%;
-  height: 60px;
   font-weight: 700;
+  height: 60px;
+  margin-bottom: 1em;
+  max-width: 90%;
+  text-align: center;
 }
