+++ src/pages/Dashboard/components/ProjectList/ProjectList.scss
@@ -1,7 +1,7 @@
 .ProjectList {
+  align-items: center;
   display: flex;
   flex-direction: row;
   flex-wrap: wrap;
   justify-content: flex-start;
-  align-items: center;
 }
