+++ src/pages/Dashboard/components/ProjectTile/ProjectTile.scss
@@ -1,21 +1,21 @@
 .ProjectTile {
-  width: 200px;
-  height: 230px;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
   align-items: center;
   border: 3px solid gainsboro;
   border-radius: 1rem;
+  display: flex;
+  flex-direction: column;
+  height: 230px;
+  justify-content: center;
   margin: 0.5rem;
   padding: 0.5rem 1rem;
   text-align: center;
+  width: 200px;
   .ProjectTile__title {
   }
 }
 .ProjectTile__button{
   background: none;
   border: none;
-  text-decoration: none;
   color: red;
+  text-decoration: none;
 }
