+++ src/pages/Home/Home.scss
@@ -10,6 +10,6 @@
 }
 
 .Home__my_projects{
-    height: 400px;
     background: #cccccc;
+    height: 400px;
 }
