+++ src/pages/Profile/index.js
@@ -1,12 +1,9 @@
 import React from 'react';
-import { useSelector } from 'react-redux';
 
 const Profile = () => {
-
     return (
         <div className={"Profile"}>
             <h3>My profile</h3>
-
         </div>
     )
 }
